+++ src/Mzinga.Viewer/ViewModels/ViewerConfigViewModel.cs
@@ -248,8 +248,6 @@ namespace Mzinga.Viewer.ViewModels
             }
         }
 
-        public static bool PlaySoundEffectsEnabled => AppInfo.IsWindows;
-
         public bool PlaySoundEffects
         {
             get